# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Crud materials Api",
        "description": "Swagger for materials crud",
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/materials": {
            "get": {
                "tags": [
                    "Materials"
                ],
                "summary": "Display a listing of all materials",
                "operationId": "fe9f481c6448e83afa2bbbcfac85ff8e",
                "responses": {
                    "200": {
                        "description": "List of materials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "All materials"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Materials"
                ],
                "summary": "Create a new material",
                "description": "Store a newly created resource in storage.",
                "operationId": "1d4d9ef681e132ef86e47298a76f6d1f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of the material",
                                        "type": "string",
                                        "example": "Material Name"
                                    },
                                    "id_type": {
                                        "description": "ID of the material type",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_env_impact": {
                                        "description": "ID of the environmental impact",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_supplier": {
                                        "description": "ID of the supplier",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "cost": {
                                        "description": "Cost of the material",
                                        "type": "string",
                                        "example": "100.00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Material created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Material created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Material"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/materials/{id}": {
            "get": {
                "tags": [
                    "Materials"
                ],
                "summary": "Display the specified material",
                "operationId": "09f24b9bcc68d549ae625cab1d222d4f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the material",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Material details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Material details"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Material not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Material not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Materials"
                ],
                "summary": "Update the specified material",
                "description": "Update the specified resource in storage.",
                "operationId": "5498fc80b227f7c483951e52cae3dd60",
                "responses": {
                    "200": {
                        "description": "Material updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Material updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Material"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Material not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Material not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Materials"
                ],
                "summary": "Remove the specified material from storage",
                "operationId": "758b0974ce93afa68e29785281c71378",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the material",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Materials deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Materials deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Materials not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Materials not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Material": {
                "title": "Material",
                "description": "Material object",
                "properties": {
                    "id": {
                        "description": "ID of the material",
                        "type": "integer"
                    },
                    "name": {
                        "description": "Name of the material",
                        "type": "string"
                    },
                    "id_type": {
                        "description": "ID of the material type",
                        "type": "integer"
                    },
                    "id_env_impact": {
                        "description": "ID of the environmental impact",
                        "type": "integer"
                    },
                    "id_supplier": {
                        "description": "ID of the supplier",
                        "type": "integer"
                    },
                    "cost": {
                        "description": "Cost of the material",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    }
}